# swagger.yaml
openapi: 3.0.4
info:
  title: Sigorta Acentesi API
  version: 1.0.0
  description: REST API for managing insurance agency operations
servers:
  - url: https://sigortaapi.elacigdem.com/api/v1
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Customer:
      type: object
      properties:
        id:
          type: integer
          description: Customer ID
        firstName:
          type: string
          description: Customer first name
        lastName:
          type: string
          description: Customer last name
        email:
          type: string
          description: Customer email
        phone:
          type: string
          description: Customer phone
        address:
          type: string
          description: Customer address
        city:
          type: string
          description: Customer city
        postalCode:
          type: string
          description: Customer postal code
        birthDate:
          type: string
          format: date
          description: Customer birth date
        identityNumber:
          type: string
          description: Customer identity number
...
    Claim:
      type: object
      properties:
        id:
          type: integer
          description: Claim ID
        policyId:
          type: integer
          description: Policy ID
        claimNumber:
          type: string
          description: Claim number
        claimDate:
          type: string
          format: date
          description: Claim date
        description:
          type: string
          description: Claim description
        damageAmount:
          type: number
          format: float
          description: Damage amount
        status:
          type: string
          description: Claim status (pending, approved, rejected, paid)
        documents:
          type: string
          description: JSON string containing document file paths
...
security:
  - bearerAuth: []
paths:
...
  /customers:
    get:
      summary: List all customers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Customer'
        '401':
          description: Unauthorized - missing or invalid token
        '500':
          description: Internal server error
    post:
      summary: Create a new customer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer'
      responses:
        '201':
          description: Customer created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  customerId:
                    type: integer
        '400':
          description: Bad request - invalid input
        '401':
          description: Unauthorized - missing or invalid token
        '500':
          description: Internal server error
  /customers/{id}:
    put:
      summary: Update a customer
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Customer ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer'
      responses:
        '200':
          description: Customer updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Bad request - invalid input
        '401':
          description: Unauthorized - missing or invalid token
        '404':
          description: Customer not found
        '500':
          description: Internal server error
    delete:
      summary: Delete a customer
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Customer ID
      responses:
        '200':
          description: Customer deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Unauthorized - missing or invalid token
        '404':
          description: Customer not found
        '500':
          description: Internal server error
...
  /claims:
    get:
      summary: List all claims
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Claim'
        '401':
          description: Unauthorized - missing or invalid token
        '500':
          description: Internal server error
    post:
      summary: Create a new claim
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Claim'
      responses:
        '201':
          description: Claim created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  claimId:
                    type: integer
        '400':
          description: Bad request - invalid input
        '401':
          description: Unauthorized - missing or invalid token
        '500':
          description: Internal server error
  /claims/{id}/status:
    patch:
      summary: Update claim status
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Claim ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [pending, approved, rejected, paid]
      responses:
        '200':
          description: Claim status updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Bad request - invalid input
        '401':
          description: Unauthorized - missing or invalid token
        '404':
          description: Claim not found
        '500':
          description: Internal server error
...